Revert "sched/idle: Reported an error when an illegal negative value is passed"
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I4QU5Z?from=project-issue
CVE: NA

--------------------------------

This patch revert ias feature from open-euler kernel.

This reverts commit fe426fa77d72338c9af7d464bd507eccf62c3f87.

+++ include/linux/kernel.h
@@ -558,7 +558,7 @@ extern int sysctl_panic_on_stackoverflow;
 
 extern bool crash_kexec_post_notifiers;
 #ifdef CONFIG_IAS_SMART_IDLE
-extern int poll_threshold_ns;
+extern unsigned long poll_threshold_ns;
 #endif
